Prepare for the tears
so I think the issue was the fact we never defined the sensor. If this was the issue, I'd like to add "WHY DOES IT NOT DEFAULT TO THE BUILT IN SENSOR?!?!" also I'm kind of tired, so that helps. This better work. I've spent too much time trying to get this stupid thing to drive.

Co-Authored-By: jojo2357 <66704796+jojo2357@users.noreply.github.com>